feat: Use a file to indicate whether the starting routine has finished or not.
The previous solution for preventing users from entering the workspace
before the starting routine has finished is to check whether the UID and
GID inside the workspace are the same as the host machine or not.
However, this implementation works only if the UID changing is the last
operation of the starting routine.

This commit makes the starting routine create a file
`/docker/start.sh.done` exactly before it finishes. This manner provides
an explicit way for the `run` script to check if the starting routine
has finished. Even if additional commands are added to the starting
routine in the future, this solution will not be broken accidentally.